move buffers and fibers to new "boss" library
